requirements.txt
asgiref==3.3.1
Django==3.1.4
django-shortuuidfield==0.1.3
pytz==2020.4
shortuuid==1.0.1
six==1.15.0
sqlparse==0.4.1
